feat: 홈 화면과 video 재생 부분 머지 후 테스트 버그 fix

+++ src/main/java/com/example/vickey/api/VideoController.java
@@ -33,7 +33,7 @@ public class VideoController {
             //1. S3에 업로드
             Map<String, Object> result = videoService.uploadVideo(file);
             String videoUrl = (String) result.get("url");
-            long duration = (Long) result.get("duration");
+            int duration = (int) result.get("duration");
 
             // 2. DB에 저장
             System.out.println("Calling saveVideo...");
